Removed duplicate coordinates from country polygons.
Some of the polygons defined as part of the USA, Canada and Antarctica
had duplicate points immediately following each other in the polygon
definitions.

Whilst this problem is insignificant and makes no difference to the
actual shape of the geometry, and in fact many programs that use
GeoJSON can overlook this problem, it is still incorrect syntax.

In particular, I noticed this problem when trying to index the data
with MongoDB's GIS "2dsphere" indexer - it threw an error exclaiming
that the exterior shell of the polygon was invalid and would not index
the field.